// src/app/operations/models/aircost.ts
import { GlobalVariables } from '../../core/models/globalvariables';

export class AirCostm {
  air_mblid: string;
  air_actual_rate: number;
  air_agent_rate: number;
  air_rebate: number;
  air_rebate_payable: string;
  air_sell_recd: number;
  air_sell_informed: number;
  air_iata_incentive: string;
  air_iata_commission: string;
  air_exworks: number;
  air_sell_info_currency: string;
  air_sell_recd_currency: string;
  air_currency: string;
  air_amount: number;
  air_mawb_chg_wt: number;
  air_remarks: string;
  air_netnet: string;
  air_iata_comm_per: number;
  air_iata_comm_amt: number;
  air_iata_inc_per: number;
  air_iata_inc_amt: number;
  air_other_chg_inc: number;
  air_other_chg_exp: number;
  air_counter_informed: number;
  air_publish_rate: number;
  air_year: number;


  rec_mode: string;
  rec_category: string;
  _globalvariables: GlobalVariables;
}